feat: a couple of refactors
